windows build regression fix: multiple compilation errors under MSVC C++17 (#115)
- processing_qwen3_5.cpp: pass std::vector<float>& instead of float*
  to resize_bilinear_to_chw()
- qwen3_5preprocessor.cpp: add static_cast<size_t>() for int32_t config
  values in ov::Shape{} brace-init to fix C2398 narrowing conversion
- qwen3_vl/vision_preprocess.cpp: add static_cast<size_t>() for int
  return values in ImageSize{} to fix C2397 narrowing conversion
- VideoPreprocesModule.cpp: replace C++20 designated initializers with
  C++17-compatible positional aggregate initialization (C7555)

+++ src/cpp/src/module_genai/modules/model/qwen3_5/qwen3_5preprocessor.cpp
@@ -235,14 +235,14 @@ Qwen3_5PreprocessorOutput Qwen3_5Preprocessor::preprocess_video(const ov::Tensor
 
     auto nchw_shape = ov::Shape{batch,
                                 grid_t,
-                                m_preprocess_config.temporal_patch_size,
+                                static_cast<size_t>(m_preprocess_config.temporal_patch_size),
                                 channel,
                                 grid_h / m_preprocess_config.merge_size,
-                                m_preprocess_config.merge_size,
-                                m_preprocess_config.patch_size,
+                                static_cast<size_t>(m_preprocess_config.merge_size),
+                                static_cast<size_t>(m_preprocess_config.patch_size),
                                 grid_w / m_preprocess_config.merge_size,
-                                m_preprocess_config.merge_size,
-                                m_preprocess_config.patch_size};
+                                static_cast<size_t>(m_preprocess_config.merge_size),
+                                static_cast<size_t>(m_preprocess_config.patch_size)};
 
     resized_video = qwen3vl_utils::ovtensor_view(resized_video, nchw_shape);
